izvorni_kod/backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.2
flask-cors==5.0.0
PyJWT==2.9.0
importlib_metadata>=8.0.0; python_version < "3.10"
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
zipp==3.23.0
# OAuth2
Authlib==1.3.0
# Firebase Admin SDK
firebase-admin==6.5.0
# Database (SQLAlchemy for production use)
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.7
psycopg2-binary==2.9.7  # PostgreSQL driver
# Environment variables
python-dotenv==1.0.1
# Email support (for notifications)
Flask-Mail==0.10.0
# HTTP requests for chatbot integration
requests==2.31.0
